requirements.txt
selene (>=2.0.0rc9)
allure-pytest (>=2.13.5,<3.0.0)
pytest (>=8.3.4,<9.0.0)
appium-python-client (>=4.5.0,<5.0.0)
requests (>=2.32.3,<3.0.0)